# api/openapi.yaml
openapi: 3.1.0
info:
  title: Booking Inventory Service
  version: 1.0.0
  description: Atomic inventory reservation API for internal order-processing systems.
  contact:
    name: Nikita Smirnov
    url: https://github.com/niksmi-lab
servers:
  - url: http://localhost:8080
    description: Local development
tags:
  - name: Operations
  - name: Inventory
  - name: Reservations

paths:
  /healthz:
    get:
      tags: [Operations]
      summary: Liveness probe
      operationId: getHealth
      responses:
        "200":
          description: Process is alive
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Status"

  /readyz:
    get:
      tags: [Operations]
      summary: Readiness probe
      description: Checks PostgreSQL connectivity and the required schema.
      operationId: getReadiness
      responses:
        "200":
          description: Service is ready
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Status"
        "503":
          description: Database or schema is unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Status"

  /metrics:
    get:
      tags: [Operations]
      summary: Prometheus metrics
      operationId: getMetrics
      security:
        - adminBearer: []
      responses:
        "200":
          description: Prometheus text exposition
          content:
            text/plain:
              schema:
                type: string
        "401":
          $ref: "#/components/responses/Unauthorized"

  /api/v1/stock/restock:
    post:
      tags: [Inventory]
      summary: Increase available inventory
      operationId: restockInventory
      security:
        - adminBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ItemsRequest"
      responses:
        "200":
          $ref: "#/components/responses/Success"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"

  /api/v1/stock/reserve:
    post:
      tags: [Reservations]
      summary: Reserve inventory for an order
      description: Repeating the same order ID and item set is idempotent.
      operationId: reserveInventory
      security:
        - serviceBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReserveRequest"
      responses:
        "200":
          $ref: "#/components/responses/Success"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
        "504":
          $ref: "#/components/responses/DependencyTimeout"

  /api/v1/stock/confirm:
    post:
      tags: [Reservations]
      summary: Confirm a pending reservation
      operationId: confirmReservation
      security:
        - serviceBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderRequest"
      responses:
        "200":
          $ref: "#/components/responses/Success"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"

  /api/v1/stock/cancel:
    post:
      tags: [Reservations]
      summary: Cancel a pending reservation and restore stock
      operationId: cancelReservation
      security:
        - serviceBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderRequest"
      responses:
        "200":
          $ref: "#/components/responses/Success"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"

components:
  securitySchemes:
    serviceBearer:
      type: http
      scheme: bearer
      description: API_KEY
    adminBearer:
      type: http
      scheme: bearer
      description: ADMIN_API_KEY

  schemas:
    CartItem:
      type: object
      additionalProperties: false
      required: [item_id, quantity]
      properties:
        item_id:
          type: string
          format: uuid
        quantity:
          type: integer
          format: int64
          minimum: 1
          maximum: 1000000000

    ItemsRequest:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          minItems: 1
          maxItems: 500
          items:
            $ref: "#/components/schemas/CartItem"

    ReserveRequest:
      type: object
      additionalProperties: false
      required: [order_id, items]
      properties:
        order_id:
          type: string
          format: uuid
        items:
          type: array
          minItems: 1
          maxItems: 500
          items:
            $ref: "#/components/schemas/CartItem"

    OrderRequest:
      type: object
      additionalProperties: false
      required: [order_id]
      properties:
        order_id:
          type: string
          format: uuid

    Success:
      type: object
      required: [status]
      properties:
        status:
          type: string
          const: success

    Status:
      type: object
      required: [status]
      properties:
        status:
          type: string

    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code:
              type: string
            message:
              type: string
            request_id:
              type: string

  responses:
    Success:
      description: Command completed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Success"
    InvalidRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing or invalid Bearer token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Reservation was not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: Insufficient stock, expired reservation or invalid state transition
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    DependencyTimeout:
      description: Database timeout
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"